+++ src/data-api/index.ts
@@ -1,0 +1,13 @@
+export type {
+  DataApiConfig,
+  DataApiOrderDirection,
+  DataApiRegistry,
+  DataApiResource,
+  DefinedDataApiResource,
+  ListResourceInput,
+  ListResourceResult,
+} from "./types.js";
+export { DEFAULT_PAGE_SIZE, MAX_OFFSET, MAX_PAGE_SIZE } from "./types.js";
+export { DataApiError, isDataApiError } from "./errors.js";
+export { defineDataApi, getResource } from "./registry.js";
+export { listResource } from "./list.js";